backend/requirements.txt
# Django Core
Django==5.0.0
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Media/File handling
Pillow==10.2.0

# Environment variables
python-decouple==3.8

# Security
django-ratelimit==4.1.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-extensions==3.2.3
